fix: move logic to service
